Logical refactor
Using some new directories to host generation process
